backend/requirements.txt
Django==4.1.13
djangorestframework==3.14.0
mongoengine==0.27.0
djongo==1.3.6
python-dotenv==1.0.0
django-cors-headers==4.3.0
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.0
